producer/requirements.txt
# Core
confluent-kafka
requests==2.31.0
python-dotenv==1.0.0
pulsar-client

